requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
icalendar>=5.0.0

# OAuth sync dependencies (optional - only needed for advanced setup)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
